applications/counter-python-nextjs-postgres/services/backend/src/pyproject.toml
[tool.poetry]
name = "Backend"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
codefly-sdk = "^0.0.6"
psycopg2 = "^2.9.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
